// blockbuzz/prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const VenueType = {
  INDOOR: 'INDOOR',
  OUTDOOR: 'OUTDOOR'
} as const

export type VenueType = (typeof VenueType)[keyof typeof VenueType]


export const InteractionType = {
  VIEW: 'VIEW',
  SAVE: 'SAVE',
  REGISTER: 'REGISTER',
  ATTENDED: 'ATTENDED'
} as const

export type InteractionType = (typeof InteractionType)[keyof typeof InteractionType]


export const RegistrationStatus = {
  REGISTERED: 'REGISTERED',
  CANCELLED: 'CANCELLED',
  ATTENDED: 'ATTENDED'
} as const

export type RegistrationStatus = (typeof RegistrationStatus)[keyof typeof RegistrationStatus]


export const VolunteerRole = {
  LEAD: 'LEAD',
  CHECK_IN: 'CHECK_IN',
  REGISTRATION: 'REGISTRATION',
  EVENT_SUPPORT: 'EVENT_SUPPORT',
  TECH_SUPPORT: 'TECH_SUPPORT',
  CROWD_MANAGEMENT: 'CROWD_MANAGEMENT',
  STAGE_SUPPORT: 'STAGE_SUPPORT',
  CONTENT_CREW: 'CONTENT_CREW',
  LOGISTICS: 'LOGISTICS',
  MEDICAL: 'MEDICAL',
  SECURITY: 'SECURITY',
  OTHER: 'OTHER'
} as const

export type VolunteerRole = (typeof VolunteerRole)[keyof typeof VolunteerRole]


export const AssignmentStatus = {
  APPLIED: 'APPLIED',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  COMPLETED: 'COMPLETED'
} as const

export type AssignmentStatus = (typeof AssignmentStatus)[keyof typeof AssignmentStatus]